refactor(api): extract sources from client.ts
